% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_ggplotRegression.R
\name{myFUN_ggplotRegression}
\alias{myFUN_ggplotRegression}
\title{A regression FUnction}
\usage{
myFUN_ggplotRegression(fit)
}
\arguments{
\item{fit}{is the before runed lm model which you want to plot now}
}
\description{
This function allows to plot a regression with all wanted values
}
\examples{
myFUN_ggplotRegression()
}
\keyword{regression}
