% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_ggscatter.R
\name{myFUN_ggscatter}
\alias{myFUN_ggscatter}
\title{ggscatter --> GLEIS}
\usage{
myFUN_ggscatter(df, xaxis, yaxis, xlab = xaxis, ylab = yaxis,
  regression = "reg.line")
}
\arguments{
\item{df}{data.frame where data is stored}

\item{xaxis}{x column of the data.frame}

\item{yaxis}{y column of the data.frame}

\item{xlab}{name of the x-axis; default = xaxis}

\item{ylab}{name of the y-axis; default = yaxis}

\item{regression}{method how the regression is done --> reg.line or loess, default = reg.line}
}
\value{
return ggscatterplot
}
\description{
plots a ggscatter
}
\examples{
myFUN_ggscatter()


}
\keyword{regression}
