% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_myGenotypeParametersControl.R
\name{myFUN_myGenotypeParametersControl}
\alias{myFUN_myGenotypeParametersControl}
\title{A GLEIS function}
\usage{
myFUN_myGenotypeParametersControl(Geno)
}
\arguments{
\item{Geno}{is a single Genotype}
}
\description{
Calculated the slope/intercept/r2 and p value of the regression curve (here in the control group)
}
\examples{
myFUN_myGenotypeParametersControl()


}
\keyword{regression}
