% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_myGenotypeParametersDrought.R
\name{myFUN_myGenotypeParametersDrought}
\alias{myFUN_myGenotypeParametersDrought}
\title{A GLEIS function}
\usage{
myFUN_myGenotypeParametersDrought(Geno)
}
\arguments{
\item{Geno}{is a single Genotype}
}
\value{
returns the above named vlaues in a table
}
\description{
Calculated the slope/ intercept/ r2 and p value of the regression curve (here in the drought treatment group)
}
\examples{
myFUN_myGenotypeParametersDrought()


}
\keyword{regression}
