% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_smoother.R
\name{myFUN_smoother}
\alias{myFUN_smoother}
\title{Smooting function --> GLEIS}
\usage{
myFUN_smoother(geno, smoothFactor, column)
}
\arguments{
\item{smoothFactor}{number; how smooth the values should get, the higher the smoother}

\item{column}{name of the column e.g. "ler" ATTENTION: WRITE IN QUOTATION MARKS}

\item{Geno}{is a single Genotype}
}
\value{
returns a dataframe with 3 columns, input (smoothFactor, rowName, smoothed values)
}
\description{
smoothing data by moving average
}
\examples{
myFUN_smoother()


}
\keyword{regression}
